Latest update

+++ src/renderer/app/components/players/players.component.ts
@@ -1,8 +1,11 @@
 import { Component, OnInit, ElementRef, ViewChild } from '@angular/core';
-import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { BsModalService } from 'ngx-bootstrap/modal';
 
 import { PlayerService } from '../../services/player.service';
 import { Player } from '../../models/player.model';
+import { RoleService } from '../../services/role.service';
+import { PlayerModalComponent, PlayerModalResult } from '../modals/player/player-modal.component';
+import { ConfirmModalComponent } from "../modals/confirm/confirm-modal.component";
 
 @Component({
   selector: 'cc-players',
@@ -11,22 +14,10 @@ import { Player } from '../../models/player.model';
 })
 export class PlayersComponent implements OnInit {
 
-  @ViewChild("confirmDeleteModal")
-  private _confirmDeleteModal: ElementRef;
-
-  @ViewChild("addPlayerModal")
-  private _addPlayerModal: ElementRef;
-
-  @ViewChild("confirmDeadModal")
-  private _confirmDeadModal: ElementRef;
-
-  public createPlayerFirstName: string = "";
-  public createPlayerLastName: string = "";
-  public createPlayerAmount: number = 1000;
-
   constructor(
-    private _playerService: PlayerService,
-    private _modalService: NgbModal
+    private readonly _playerService: PlayerService,
+    private readonly _roleService: RoleService,
+    private readonly _modalService: BsModalService
   ) { }
 
   ngOnInit() {
@@ -37,28 +28,67 @@ export class PlayersComponent implements OnInit {
   }
 
   public createPlayer(): void {
-    this._modalService.open(this._addPlayerModal).result.then(() => {
-      this._playerService.createPlayer(this.createPlayerFirstName, this.createPlayerLastName, this.createPlayerAmount);
-      this.clearCreatePlayerInputs();
-    }, () => {
-      this.clearCreatePlayerInputs();
-    }); 
+    this._modalService.show(PlayerModalComponent).content.result.subscribe(
+      (player: PlayerModalResult) => {
+        const newPlayer = this._playerService.createPlayer(player.firstName, player.lastName, player.startingAmount);
+
+        if (!!player.role) {
+          this._roleService.assignRole(player.role.name, newPlayer);
+        }
+      });
   }
 
-  private clearCreatePlayerInputs(): void {
-    this.createPlayerFirstName = "";
-    this.createPlayerLastName = "";
+  public editPlayer(player: Player): void {
+    this._modalService.show(PlayerModalComponent, {
+      initialState: {
+        edit: true,
+        firstName: player.firstName,
+        lastName: player.lastName,
+        role: player.role
+      }
+    }).content.result.subscribe(
+      (updatedPlayer: PlayerModalResult) => {
+        player.firstName = updatedPlayer.firstName;
+        player.lastName = updatedPlayer.lastName;
+
+        if (!updatedPlayer.role) {
+          this._roleService.unassignPlayer(player);
+        }
+        else if (player.role !== updatedPlayer.role) {
+          this._roleService.assignRole(updatedPlayer.role.name, player);
+        }
+      });
   }
 
   public deletePlayer(player: Player) {
-    this._modalService.open(this._confirmDeleteModal).result.then(() => {
-      this._playerService.deletePlayer(player);
-    }, () => {});
+    this._modalService.show(ConfirmModalComponent, {
+      initialState: {
+        title: "Delete Player",
+        text: `Are you sure you wish to delete ${player.name}?`
+      }
+    }).content.result.subscribe(
+      (result: boolean) => {
+        if (!result) return;
+        this._playerService.deletePlayer(player);
+        this._roleService.unassignPlayer(player);
+    });
   }
 
   public setDead(player: Player) {
-    this._modalService.open(this._confirmDeadModal).result.then(() => {
-      player.setDead();
-    }, () => {});
+    this._modalService.show(ConfirmModalComponent, {
+      initialState: {
+        title: "Player Dead",
+        text: `Are you sure you wish to set ${player.name} to dead?`
+      }
+    }).content.result.subscribe(
+      (result: boolean) => {
+        if (!result) return;
+        player.setDead();
+    });
+  }
+
+  public getPlayerRole(player: Player): string {
+    const role = this._roleService.getRoleForPlayer(player);
+    return !!role ? role.name : "";
   }
 }
